readme update,cleanup,ediff-fixed-HLing
@@ -1453,13 +1453,6 @@ added-pkgs
- removed declutter not using will archive

@@ -1468,6 +1461,9 @@ added-pkgs
- This is new pkg for permanent-scratch-buffer
- This was needed, luckly had backup.
- not sure why I have lost this Doom/scratch-buffer repeatedly.
- started doing my own back ups, this saved me just recently.
so I am hoping this makes it easier and automate it. hoping? will see.

@@ -1746,7 +1742,7 @@ my-keybindings
- fixed typo that broke keybinding.

@@ -2316,8 +2312,8 @@ elfeed-org
- removed rss feeds from repo. will back local

@@ -3070,15 +3066,16 @@ media-note
- another typo ???

@@ -3094,9 +3091,18 @@ Ediff
- Felt with out numbers or fringe the highlighting in ediff just did not
work. Could not tell were the cursor was or what hunk it was on. So this
makes it functional now may add a "even" noticed that after moving one I
felt like the color should indicate that.