clean-up,reformat, purge and refine
- mixed bag of changes some purely for aesthetics
some needed purging while some needed to be formatted proper.

@@ -9,10 +9,10 @@
@@ -153,7 +153,7 @@ dashboard
- aesthetics

@@ -193,12 +193,11 @@ dashboard
- indent was wrong and the last line of the function seems to be left
empty so got rid of it.

@@ -590,15 +589,22 @@ evil snipe
- this first part is an unknown. But the reason for the change is
flycheck throws a error message if I wrap the prefix in a con with
descriptor. Has no prob with just the key so purge till know more.
- the second part has to do with me making a decision on what prefix I
wanted for evil-easymotion. Like the convenience of the space bar, but
did not like that it was squashing others I have not tried. Wanted to
explore some of the others so the Solution was:
- To use the later and add the extra bindings I liked from the previous
back. This worked perfect I got what I liked about the two in one.

@@ -1462,8 +1468,6 @@ added-pkgs
- Did not like this pkg. It did everything through customs-menu and I
dispise this menu as well as I had no idea what this was actually doing
I have a permanent-scrach-buffer and thought this would make copying to
a backup easier but it did not seem to be doing this ...purge it then.

@@ -1898,6 +1903,15 @@ mpv.el
- This got erased long time ago because a seperate pkg had the same
setting so I deleted this one and now in after some time I am thinking
that I should keep this one just in case that pkg gets uninstalled and
the function go with it, so this was added to medigate against this
happening.

@@ -2069,16 +2083,6 @@ deft
- This was left behind after a reformatting and was there just in case.
It appears my new settings seem to be working so I will clean this up.

@@ -2138,19 +2142,19 @@ Elfeed
@@ -2169,8 +2173,9 @@ Elfeed
@@ -2181,7 +2186,7 @@ Elfeed
- this change comes from using the doom config as my own to get access
to the use-package function and know what is preset already so if I want
to make a change I can know what is already set.
- This part I needed to add to my settings but I had two issues with it,
one was I was curious about the fact that he used a ":after elfeed-show"
and not the just plain after elfeed that I had.
Decided to go with what he had and replace what I was using.
- The second thing I needed to address was the fact that one of these
did not work the way I expected it to and It the way it is described so
I changed it to something that I believe will work better. But now that
I have made these changes I can get rid of the backup that was left
behind just in case it did not go as planned.

@@ -2392,90 +2397,90 @@ Elfeed-summary
- This is interesting cause I did not do this last change.
- I am thinking I did change this but long time ago and it has been
pooched like this ever since and something ether doom or the package it
self had corrected some Indentation problems on top of that there is a
extra part that has been added so I would imagine the package updated
then formatted it properly.

@@ -3098,10 +3104,12 @@ ediff
- added missing braces and a setting for "even" have yet to try so could
change.