added a function to make video-notes easier
- quick update this change has not proved it self it was an idea that I
thought I would want in the future.

@@ -1406,6 +1400,19 @@
- this function here will give me a interactive toggle of background
transparency.
- the reason was my first try of a package I load a while back. It is
for taking media-notes hence the name. It has a toggle on top feature
which got me thinking about one of the limitations of my homemade
version of this was splitting the screen, made grabbing something
written on screen hard. I suspect maybe that is what the screen shot
tool is good for but have yet to try it out. what I did feel like is
having a screen that was transparent would work or should work. I use a
similar thing for my terminal and it has been useful so I am sure I
could use it for this.
- have older version of a similar function just not the interactive part
it was set but could be toggled, did not use it after getting use to
emacs and like having code block a darker color. This looks like a zebra
if you try to use transparency so archived it. This is the new and
improved one, made with the built-in function. Only have one use but
maybe of help. will need to spend some time with it to make that
decision.