added calendar settings, rest mostly reordering
- most will be clean up, or re-ordering, and of course purging

@@ -2,7 +2,7 @@
- added some subheaders and would like them hide to begin with

@@ -111,14 +111,13 @@ startup-settings
- removed some redundancies
- this other part is a problem still trying to figure out the best
method for solving. The problem is my permanent-scratch-buffer is not so
permanent. this like the third or fourth time I have lost everything in
this buffer. real pain.
- this is one of the suggestions I got from an past issue that was fixed
I believe, but in the comments it is suggested that what might work
better is to use a hook rather that trying to change the mode. But it
looked like they added the variable then to set a default mode just he
used 'setq-default' which I did not. Not sure if that solves the problem
or not. Might have to revisit this in the future and will be pissed
cause it lost all the info that was on there. will have to see if that
happens or not.

@@ -137,13 +136,20 @@ startup-settings
- here there is two things happening first is to quit the noise from
flycheck-mode in org-blocks. Not sure if this is working yet. will have
to check back with this.
- the second thing is back ups, this goes back to losing all the data in
the permanent scratch buffer. This is the first steps to having that
file backed up and have more than one. The thing here and this really
pisses me off is the fact that I was already backing this file up.
But it was with rsync and although it will do the job it does not keep
copies like git. So what happened is the scratch buffer got erased, I
then did a backup of the empty file now that did not help me when I
finally clued in that it was gone. least with some kind of history can I
see this being prevented so that is what the last part of this is.

@@ -348,13 +354,6 @@ org-settings
@@ -462,7 +461,7 @@ org-capture
@@ -523,7 +522,6 @@ org-journal
- this is just cleaning things up

@@ -597,6 +595,105 @@ calendar
- added a calendar setting robbed it from doom-configs

@@ -1483,26 +1580,6 @@ personal-functions
- This was throwing an error when I tried using it so since I have
multiple functions to open a empty-buffer that instead of fucking with
it. It was easier just to purge and forget.

@@ -1832,6 +1909,8 @@ my-keybindings
keep having a hard time remembering this commands Keybindings so I made
another to see if that works better.

@@ -2003,30 +2082,27 @@ url-handlers
- This was a mess from the start I think I found the function in some
gist that I don't remember. It was a hackey in the way it was built and
I kept removing cruff till I broke it and needed to redo it but this
time use something that seemed more reasonable.
- still figuring out why this function start process does not seem to
work me?

@@ -2144,115 +2220,8 @@ elfeed
- added a function section and placed all elfeed functions in their own
block under a subheading this make moving and finding easier.

@@ -2440,13 +2501,12 @@ elfeed-curate
- this is another example of things that should be easy but end up being
a bain of your existence. a real pain in the ass cause the belief that
it should be as simple as making a hook, not! It is never that simple
the hard complex things go off with out a hitch the simple things now
there you got a problem and this is one of them
- the problem isnt a problem as much as I don't see the reason to
constantly needing down the 3 lines to the first entry, It should be
able to do this with out to much trouble, but oh no that is not going to
happen here in emacs land. No we want to make all the simple things the
hard things and the complex should be easy.
- this is the left overs of multiple tries at a simple move 3 fucking
lines hook that never worked.