lot of changes to old settings added Doom corfu/cape
- cleanup of some old settings that I have had since the start.
- In the process of switching to using a modified Doom config
- left both copies to document the switch. from old to new

@@ -91,11 +91,6 @@ startup-settings
@@ -106,8 +101,6 @@ startup-settings
- pruge old settings that are not needed

@@ -116,8 +109,7 @@ startup-settings
- these persistent scratches keep disappearing this did not help so
undid

@@ -143,13 +137,6 @@
- read that it would be a good idea to set back up in their own folder
- realized I don't have backups being created so I turned it on thinking
this would help. But started see weird behavior when trying to edit a
autoload function. So I am shuting this down.

@@ -1020,7 +1006,113 @@ corfu
@@ -1098,6 +1190,64 @@ cape
- these are the new corfu/cape settings I was going to show the old but
I commented the header not knowing that it makes it not show up in the
diff. oh well

@@ -1580,6 +1730,29 @@ personal-functions
- re-installed this after figuring out what I had broke on it. noticed
it missing so must use it now and again.

@@ -1884,7 +2057,9 @@ my-keybindings
- trying out eshell thought this would help

@@ -2082,17 +2257,19 @@ url-handlers
- This was a real boner on my part, this was working fine till I felt
the need to "fix it" and what I really did was "fucked it".
- this was another one of those times when I thought maybe that backup
is working against me and that why it seems to never change even thou I
have changed the function many. ??? still not sure undid backup to be
safe.

@@ -2106,8 +2283,8 @@ Rant
- this is just a rant this whole customize thing with the fuck menu
dislike it with a passion. try to avoid it but forget when something is
set there. certain things won't change until I go in to that menu and
hit the set and save button. really can fuck you up if you forget that
is the case with the function you might be monkeying with. This is an
example could not figure out what I was doing wrong here cause it just
kept giving same error. Not till after did I remember that it was a
customized function and I needed to change it there as well to get any
of the settings to stick. fuck!

@@ -2222,12 +2399,16 @@ elfeed-functions
- This solved a problem and I feel stupid that I did not see this to
begin with. But any way this got rid of the "async-shell-command-buffer"
which started to become a pain to have to work around in elfeed since
the buffer stays there and you have to remove it your self. This got old
fast. But spent all the time trying to get rid of the buffer when I just
needed to change commands ...Duh.