added my scratch file, purged old corfu/cape
- going to add my permanent-scratch which seems not so permanent.
- added a symlink and left most other things so the settings should be
  default

@@ -2255,29 +2112,31 @@ web-url-handlers
- this has given me quit a bit of trouble as of late and I can only
contribute it to user-error .... that would be me fucking with things
that were working fine till I decided to "fix them" or my favorite
"cleaning up code" that one you know will bite you in the ass, sooner or
later. So long story longer. I am trying to revert this back to when it
use to work. then will decide what has fucked this. Was it me? I just
did a series of updates to everything. So the only I can think of
telling if one or the other is at fault I will need to take this back to
when it worked and see what happens then. Maybe something has changed
and I am unaware and just assume it was me. chances are it was me. But
this should help in figuring that out.