update to configuration
-
@@ -108,8 +108,6 @@ startup-settings
- This I believe was causing the scratch to restart but blank losing
data so I will remove it and keep it under git to prevent in future.

@@ -432,51 +430,64 @@ org-capture-templates
- May have solved the bane of my emacs existence. The costume variable.
Have been trying since the beginning of my emacs experience trying to
get customized settings to be set from the config. I believe I finally
figured this stupid little thing out. messed around with
org-capture-templates for ages trying to get it to set. Not till now did
I realize that all I had to do was hit the revert button and then it
would be set by the config. This whole time I have been eval, then go to
menu , then hit the save button, just to get it to change. Finally.
Knew it would be something small and right in font of my face but could
not see it.