Fixed "web-url-handler" block
- This may have something to do with using autoloads and
make-backup-files.
- these functions got a "fix" or a "clean up code". under normal state
they would have failed right after the changes made to them. But instead
acted like the change was working fine when in fact the changes had
broken the function. But ether the backed-up's or the auto-load.
Prevented these from being changed. As a result, I was caught off guard
when they all of a sudden stopped working. could not come up with reasons
on what could have happened to it, since I had not changed anything for
a while in that setting.
- the fix for the problem was to revert everything back to were they
worked. remove (make-backup-files) as well as the ;;;autoload.
- next was to go to the customize-menu and revert any thing that was set
there. Will need to check if changes that were made are to a
custom-variable or not.

@@ -2136,12 +2134,14 @@ web-url-handlers
- this was the fixed version

@@ -2156,20 +2156,6 @@
- Have no idea why I had this kicking around still this has not been in
use since the first day I put it in there years ago. how add the hell
did I miss this for this long.
- I installed this right off the start, when I had no idea what I was
really doing. For some reason I never thought of removing it and have
no reason as to why. MY BAD.