got URL handlers working again, small fixes
- mainly was able to get the "URL" handlers working again after some
changes.
- small other changes

@@ -174,7 +174,7 @@ calendar
- Here is were I solved the problem why I could not see any appointments
in the calendar. Turns out I was using the wrong command. So now it
should show the sources that I have listed in the config. fixed

@@ -188,10 +188,8 @@ dashboard
- I assumed that this was just for emacsclient but after commenting it,
I found out that no it is part of the dashboard startup so add to the
function.

@@ -1896,7 +1894,7 @@ my-keybindings
@@ -1913,14 +1911,14 @@ my-keybindings
- first keybinding is for the new calendar
- second is for debugging reasons, this caused a bit of confusion on my
part thought there maybe two functions with the same name, turned out
just to be a keybinding and that was all. but it was using the same
function I was working on so, commented as not to have any problems.

@@ -2127,32 +2125,27 @@ web-url-handlers
- this is what the final version looks like. archived the other
function.