added journalctl, strip tables, info nav
- added journalctl-mode, I have used this but needed terminal. This
allows me to use a emacs buffer way easier.
- added a found function that will strip a org-table from data tables
- adding a package by Charles Choi that is better defaults for
navigating the help, info, Docs in general.
- have yet to do anything other than download the file to my repo will
look at later. future project.
there are mixed in changes that are aesthetic, or clean up.

@@ -1613,6 +1622,19 @@ personal-functions
- this is a function I felt I needed after dealing with a copy paste
that did not work well. needed to un-table a table and this is what I
found and thought I would keep it.