added full screen time and repeat for headers
- added a display time in full screen, found full screen and liked it
but it hides the time so made the time show in mode line just when in
full screen
- added a repeat to head jumping. Not as use full as thought since a tab
will break the cycle and will need to be done again which is more inline
with what I wanted. Will keep it for now, it still works and doesn't
seem to be that expensive.