added casual-info but have not set it in config
- added casual-info to packages but have not set it. the reasons are as
follows.
- there are keybinding which is all I may need and found after
installing, so not sure if rewrite the keys would be a easier thing to
do still need to decide on a key to use for it. Not sure how expensive
it gets after setting it so still in the discovery phase will need to
make a call one way or the other.