added xml-hide may not keep it but try it out
- added xml-hide thinking this would help in getting copies of the
keybindings that I am using and have a record of it so it can be checked
with out needing to open the app.
- the problem is they are kept in xml files and would like just the text
from these files and not the elements. thought this would do that but
after messing around trying to get a local-repo package to work I
realize that it was not going to work on the files I wanted them to work
on. They would remove everything and that was no good. could mess around
with the regex to fine tune it but it sounds like I am trying to
re-invent the wheel here and there is a way of doing this with little
effort I just don't know about.
- will a least have this record of doing it
- trying to stage one thing at a time and see if it is better.