xml-removed
- This was more about what needed to be done to get a package from
myrepo to load. there are many possible ways. was trying to narrow down
the least that was needed.
- one problem found was defer t was causing problems so in future don't
use right off the bat. took a while to debug what was wrong. After
several changes made but no progress, that was when I noticed the defer
was the problem. Was able to get it to work with out the package.el
setting which is confusing because other times I found it was this that
was preventing it from loading. So still unsure about what role the
package.el plays.
- this was only 68 lines of code or something so it seem like something
easy to learn stuff with. Like what works and what does not. Still
confused but learning little bit more. may come back to this in the
future. So it has been archived for this reason.