(add)ready-player,subdiv dired,re-do dired-preview
- added package ready-player, will give it a try.
adds a preview panel with info and thumbnail for media-files.
- add the declaration in the dired section. I noticed by adding this
this section was big enough to subdivide into smaller section to nav.
- subdividing helps moving in and around in the section as well as
editing is done in a more narrowed buffer. So convenience is the reason.
- this caused several problems one of which I am still work on but it
can be worked around for now. One of the things I noticed is no were
does it say that dired-preview is needed but. regardless I had to
do it's settings all again in order to get them to work together.
- default way to install both was to install right at load time and kick
both of there modes in. this added to load time and I thought this could
be improved. So after trying to set both of these in all the different
ways that were possible. Any of the ways I tried to defer did not work.
I felt like I did not like have a preview pop up always was good some
times I go into a directory to search it and find the preview in the
way.
- one of the first problems was it would not update as I moved. After
enough searching I found the settings that was causing it to hang.
- it is in the custom-file diff included. two settings that when on.
wants confirmation before it kills any dired files. this needs to be nil.
- loading ready-player-mode right off the start was what adds the load
time but it needs dired-preview on to work and I decided to just toggle
it instead so did not need both of these to be loaded and global-modes
turned on. two keybindings seemed too much should be easier.
- tried using the :after keyword but could not get it to work.
- what finally worked was using the :hook keyword. toggle preview-mode
would load ready-player and start it's mode.
- the third problem is it won't clean up the preview buffers after
quiting dired it should kill all the other buffers that you visit but my
are always still there when I try to kill emacs or look at the buffer
list every preview is an open buffer. still have this and will see what
can be done about it but I can use kill all buffer key when this
happens.

custom.el @@ -120,6 +120,8 @@
this is the setting that needs to be nil