(change) keybindings debug an issue
- changed these keybindings because they were using the (C-c C-c) prefix
and this combo may be causing havoc, would get an error when using it in
the journal to exit it. then I noticed that something was resetting my
keyboard settings for capslock to control. So will try to eliminate them
and see and so far I would say it did make a diff. no errors and
keyboard settings seem to be keeping intact. will need more time to
confirm.