adding evil-embrace
- added evil-embrace refused at first until I had better understanding
of evil-surround. Used several times and found that I could remember to
use text-objects to wrap something but then had to think to hard on how
to change a surrounding and would forget. So I think it is safe to say I
gave it a chance and now I am willing to look at different options.