dashboard items changed
- not using bookmarks, thought I would try some others to see if they
are of more use.
- nothing wrong with the bookmarks it just everytime I will use the
keybinding instead. I have a good Idea of what those bookmarks are.
- problem #1 became after adding agenda it will not show anything from
my personal file, I seem them when something breaks and they show up
which is telling me they are there something is stomping on it and
resetting everything to a blank agenda.
- while messing about I find that by un-filtering the agenda I get
anything that changes in the agenda files. This seem to accomplish what
I wanted from projects, basically show me what I was doing last, what
was I working on? stuff like this would be more useful. This should give
me anything I grabbed or thought about before it becomes a project so I
will try this instead.