toggle flycheck in src code editing
- The overly noisy warnings in src-block-edit, become annoying right
away.
- starting to think this is not that useful to me. Starting to seem to
be overkill to have this turned on globally. I spend most of my time in
the .org file add use org-edit-src-code when needed so it's not working
well in either .... so why do I have it running all the time. Might try
out fly make and then just toggle one on when needed.