(add) helpful functions for the Doom module dir.
- surpise function ;; Note try harder to record were these come from and
whom. This was something grabbed from a random config that I went
through but did not record the who or were. It was grabbed from.
- So while zipping through some config file that was posted on reddit. I
grabbed this snippet. It was not exactly what I needed but it had most
of the parts. So grabbed it and changed it to what I wanted it for. Did
not take the name down because at the time I did not know if I was going
to use it or not.
- the surpise was it was more complex then I thought by just reading it.
After I ran it once I noticed it was only searching the org files which
are the readmes. So this first consult search will search just the
readmes, then the other version that I made will search the whole
directory.
- So nice. Better than I thought. This was something that was in the
back of my head and was searching for but did not come across it. That
was a consult search that search the directory recursive. But the
default search would do this it would not show any previews. So there
was no context around the search. but knew this was possible by other
searches I had done. This looked to be that. found it to be more with
the extra flags. So long story over.