function name change +pkg org-rich-yank
- magit grouped these together but should be separate.
- top is a function rename because when it come time to find this I
- could not remember anything about it, or what keyword to use to search
- for it. So it needs to not have such a generic name.

- the second thing is something I have installed and un-installed
several times this will be the third time giving it a try. That thing is
org-rich-yank it is simple yet hard to grasp and I think it has to do
with the wording around this. Anyway it just paste code with the
appropriate org blocks attached and a link to were it come from. Sounds
good and something I might use. The confusion happens the min you put in
the settings. Its called yank but it has more to do with the paste side
of things and I think that is were it get confused. first the key is in
org-mode-map ??? but it's called yank??? and it picks the language from
source??? can you see were it can be confused with something else. If it
can only be used in org-mode-map how does it read the source language if
it only works in org. Makes no sense. It's the yank in the name that is
deceiving. When it actually works when you do the paste. The paste
happens in an org-file.
- on top of that is the fact that it demands to be loaded right off the
hop. Which can be seen in the load time. So if it confuses you off the
start then adds to your load time everytime. You just end up getting rid
of it rather than figuring it out.
- But third time I understand it better than I did before and got it to
work with only minor load time change.