add (+config) this is my own package for config
- There is a long story to go with this, but the point is I added this
back to my config from the archives.
- It is a goup of functions package in one file. I made this on my own,
and was rather proud of my accomplishment. It is not a basic concept and
more of an advance thing and I was able to do it. And it work well.
- I then felt like there probably a reason for this not have been
already made and used. I then thought I would say "I did that" and
archived it. but never forgot about it. Just want to see if there was a
good enough reason to have it. Or did I just put together something that
is not that useful and that is why it's not already there.
- Now I went and dusted it off and it still works and I thought you know
I might want to use this. The idea is simple.
- there are times when I know the config is open on one of my work
stations and maybe more. but not sure which one. This should fix that by
having a dedicated named workspace. So there now if I want I can open
the config in it's own workspace and know were that is.