Will switch out Flycheck for Flymake
- Not convinced that I am getting a lot out of Flycheck. So will try the
built in one sounds better since Flycheck is starting to annoy me in
editing source blocks were I could actually use it and it just fill the
page up with warnings that are not real errors. when looked further
sounds like this has been like this for a while and every one just says
it not their problem and it just stays like that. So at this point all I
want from it is to show errors thats it.