set popup rules for Hnreader and reddigg buffers.
- this took some time because I had to go down the window,buffer,frame
rabbit hole to understand what was going on. just so I could come up
with a way of controlling these two buffers. It's a tangled mess. Hard
to wrap your head around, but what added to this is the fact that doom
has it's own buffer management system going on. This just added another
layer of complexity to the whole thing. Wondered why he would do this. I
realized I had did it by adding a module at the very start but had no
idea what it actually did. The description of it made it sound like yes
this is something I want. Not understanding what was there and what this
was doing. So a hell of a learning curve there. This is why it took so
long to come up with something. This is not what I started out to do but
as I was trying some of the many settings I came across this version and
I thought it worked better than the one I had in mind. So knowledge has
been learned through the process of trial and error.