(easy-motion) complete redo,add transient to a pkg
- easy motion gets a complete redo. while doing this I discovered and
use Doom-Emacs settings.
- build by myself from the ground up a transient for it with 30 some
commands. It wasn't till half way through it did I notice I did not need
have of them. But made a transient that worked and did it all from
scratch.
- then realize I had downloaded a package in myrepo that I had plans to
make into a file to hold all transients. So that is what I did and it
worked just like I thought it would. Ran into I small wrinkle, the next
day I lost all my keybindings. So needed to fix this and it happened to
be a function not a bug. He had it set not to use the keybindings that
were there. So now it works and I think works better than before.
- what needs to happen is I will need to use it more. So I get use to it
then get familiar with other parts.
- So the main meat and potatoes is in the pkg:"tray" .
- Was originally made by Homepage: https://github.com/tarsius/tray
- I got rid of two thirds of what he had in there and added my own so I
no longer see it as his. His is only the skeleton, I plan on putting
others here as well. That way my config isnt full of them. I know where
they are. So all is well, that ends well. It work well and am happy with
the results.