removed search functions I installed for testing.
- This started off by noticing and wandering why I did not get previews
in the default searches I was doing.
- I already came across the settings that remapped damn near every
search to a consult search. This sounded great but were are the
previews?
- I thought the whole thing that made the search better was the
previews. So where are they I could use with them at times. So then I
realized. Ya that is what is missing.
- so looked around and found and tried some other functions that were
consult searches and they had the previews like they were suppose to. So
this got me to look closer at some of the comments. That is were I came
across the explanation about it. It then made sense.
- In the comments is were I found out he turns the previews off because
it can cause slow downs in large files and I could see how that would
happen. So that made sense. That is were I found out it is set up to
toggle the previews, when you need them. This was helpful in making the
default searches much better and more in line with what I thought should
be there. So I no longer need these, they helped me in figuring out this
problem. But now I solved the problem I can remove these.