added to the "tray" pkg add vertico-menu transient
- Put some work into this, because I wanted to make one for vertico for
a while. So made it tested most of it seems to work, had to fix
keybindings. "keymap-set" is a different and does not play well with
others. So could not just use a Doom-macro, to accomplish it. Needed to
adhere to some rules, Some of which sounds like it was mac os related.
In the info on it, it uses terms like hyper, alt, not just, meta but
both meta and alt, this would make me believe who ever made this likes
to put those in everything he does, just to make it more mac os centric.
- whatever?