added a new package "substitute" for convenience.
- Have known about this package since he first brought it out. I even
downloaded it to my local repo then.
- the problem was I was not familiar with the default way Emacs did
this. I felt it was important to figure that out first before using the
package. Now I am more familiar with it so I finally installed this to
see what it does and it just makes doing it in less keypresses. Which I
like but not needed unless you are doing lots of this. Me not so much
but I still like it and spent some time getting it to work with a lazy
load, so will keep it for a while. Works fine