added (lookup) to tray transient. latest project.
- this was a bit more work than anticipated, a lot of references needed
to cross reference with keybindings. Filter out possible candidates.
Lots there to unpack once you start. So rather proud of this one and I
think it was not done in vain. I think it could be useful and used more
than once. Since I tried to use the original keybindings, I should be
able to use those with out no changing anything so some of these should
work the same with or without a transient.