(tray)(wiki-summary)(axy) added packages.
- added these three packages, axy,tray,and wiki-summary.
two of these are being redone while axy is the new package.
In trying to get the new package to load. I noticed the gap I had in
how to load a package from a local-repo. Did some research and decided
I was doing it wrong and changed the way I was loading them. So what I
did was load the directory to path. Then was able to use use-package! to
load it when I thought it should be. This dropped load time so and
everything is working so far. So this must be better than what I had.
- Now have more control over how local-packages are loaded able to defer
them as needed.