gave (axy) the axe.
- This package was important in the fact, that it was not about the
content of the package. But what I used the package to do. That was
figure out how to properly load a package from a local-repo and have
some control over how it loads. So from that stand point this was an
helpful package. But after I accomplished that I had got to try the
package itself. It did little I could not do already in a sense. It felt
awkward and hard to get rid of with out copying it to leave the buffer.
Could accomplish the same with what is already there. It did not make
playing around with snippets any easy. Just my opinion. Archived for
future reference. Maybe things will change after a bit. but for now I
don't feel this is something I am in need of.