update (corfu) merged some new settings.
- Even thou I was trying to sit out updating. I was hoping to wait till
the dust settled, then update. But I made the mistake of installing a
package that caused a glitch in a local file not updating and it
refusing to compile so no sync. Long story longer I needed to force an
update.
- after said update there was couple of things that had stopped working
or were working but in a weird way. Corfu's case was no [tab] complete.
- It would not grab the next candidate. after reverting back what I
could (:pin) I looked in issues and found what looked to be the problem
but I could not get this new patch just by updating. So I looked at the
diff. and changed the things that needed to be changed manually.
- So what I am saying is I modified the a modules config file in order
to get things to work.
- On top of all that I needed to update my config to his new changes. So
that is what this is. All the stuff I could see that had changed from
earlier version.
- So lots changed but I can say it is all working now and I suspect that
what I changed will be changed officially after Emacs 30 drops.