[my/new-named-note-function] made a new function.
- will go find were I got the original function from but it needed some
tweaking. But I have made it mine. I like it and find it useful.
- what it does is, ask for a title of the new note then make a file name
from the title in my notes directory, so all I do is save it after.
Nothing more needs to be done. So it matter not were I am it gets saved
to the right directory with little intervention.
- pumped about this. Even made a whole project file for it.
- May expand the idea to a transient for notes. The reason for the
project file. This could turn into a few more projects. Will see.
*note the comment or uncomment line function is the result of sorting.