[pomidor] gets transient and added options.
- tried to do a trial and error with different ways to load a transient
but all was lost, I had made the mistake of leaving part of the template
in the function and that is why it was not working. So did not realize
till I had it in this format. After fix, it worked, so no use in
replicating the test all over again. Will leave like this.
- made the transient cause I would forget the keybindings, but now after
messing with this that long I think I know the keybindings now and this
has already become redundant. Oh well, I learned a thing or two so still
worth it.