(ace window) added
- I needed time to get use to some of the default window selecting to
see that I still need this. Don't have more than two windows open at one
time if I do I usually open a new workspace. But now I have a use for
this, there have been times were I have had 3 windows open and found
that if one of the windows is split it can take a bit to get to it. This
makes that easier.