[literate file] moved to projects.
- Lots to be said about this. This is my attempt to use an org file to
manage a long custom variable.
- First reason for moving is because it would tell me in the echo area
that it was trying to tangle with config.org??? Did not understand.
After reading some of the docs it appears that yes any org file in Doom
Directory is auto-tangled. Did not know this. So moved to projects and
use it from there, just put a absolute path in the tangled HEADER-ARGS.
- Second is the fact that the minute I started fucking with this it is
like some decided to rebuild elfeed-summary??? Not sure why? Looked at
the repo and no updates have happened in months. So it is strange that I
have a pet project that I don't usually mess with. But the minute I did
it needed to be rebuild in an update and from that point it had a
problem so now it feels like what is not working must of come from my
intervention but I am thinking that what ever the problem is it has to
do with something I had nothing to do with. Just seems to convenient
that it all happened the way it did.
- Something seems fishy here. Like I am made to believe that it must be
something I did that is causing the new problem but I am now thinking
this has nothing to do with me and could be a dependent that has changed
and that is why it all of a sudden needed to be rebuild. Why this
happened just as I start messing with it is a little sus.