[elsum] tangle project. changes explained.
- This change to lexical-binding line is from flymake complaining that
this is the proper syntax. So there. The thing is I was reading some
where that custom set variables turn off lexical-binding I am not sure
about any of this will leave at this for now.
- The code change is what I believe it should be. But if you let Emacs
format the first group it will use the indent that I am replacing here.
- tried both ways and it did not seem to make any diff. So will use the
one that makes more sense to me.